+++ pdns/test-statbag_cc.cc
@@ -41,9 +41,9 @@ BOOST_AUTO_TEST_CASE(test_StatBagBasic) {
   s.declare("b", "description");
   s.declare("c", "description");
   s.inc("a");
-  BOOST_CHECK_EQUAL(s.read("a"), 1);
+  BOOST_CHECK_EQUAL(s.read("a"), 1UL);
   
-  int n;
+  unsigned long n;
   for(n=0; n < 1000000; ++n)
     s.inc("b");
 
@@ -83,7 +83,7 @@ BOOST_AUTO_TEST_CASE(test_StatBagBasic) {
 
 #ifdef UINTPTR_MAX  
 #if UINTPTR_MAX > 0xffffffffULL
-    BOOST_CHECK_EQUAL(sizeof(AtomicCounterInner), 8);
+    BOOST_CHECK_EQUAL(sizeof(AtomicCounterInner), 8U);
     s.set("c", 1ULL<<33);
     BOOST_CHECK_EQUAL(s.read("c"), (1ULL<<33) );
     s.inc("c");
@@ -92,14 +92,14 @@ BOOST_AUTO_TEST_CASE(test_StatBagBasic) {
     s.set("c", ~0ULL);
     BOOST_CHECK_EQUAL(s.read("c"), 0xffffffffffffffffULL );
     s.inc("c");
-    BOOST_CHECK_EQUAL(s.read("c"), 0 );
+    BOOST_CHECK_EQUAL(s.read("c"), 0UL );
 #else
-    BOOST_CHECK_EQUAL(sizeof(AtomicCounterInner), 4);
+    BOOST_CHECK_EQUAL(sizeof(AtomicCounterInner), 4U);
     BOOST_CHECK_EQUAL(~0UL, 0xffffffffUL);
     s.set("c", ~0UL);
     BOOST_CHECK_EQUAL(s.read("c"), 0xffffffffUL );
     s.inc("c");
-    BOOST_CHECK_EQUAL(s.read("c"), 0 );
+    BOOST_CHECK_EQUAL(s.read("c"), 0UL );
 #endif
 #endif
 }
